Temporary php to retrofit data
Two temporary php files are included to retrofit Player data to be
compatible with the prior two commits

commit 149376a7e134862514b0773d76d76b6fc2a2f3dd
commit 567fb0ba281680034861a68514e0a0e5c587c18a

Each of these files need only to be run once to retrofit the data.

These files can both be deleted from the live qwikgame system and
from this git repository once the retrofit is complete.