Implement qwik.php spliceOrb() to replace extendOrb()
qwik.php function parity() estimates parity by computing the parity
from both the player and rival perspective and taking the average.
This approach is problematic when there is significant disparity
between the two results because the relative reliability of the two
parity computations is not taken into account.

qwik.php function spliceOrb() takes the alternate approach of
splicing the Rival Orb onto the Player Orb and then using the
spliced Orb to compute the parity (which automatically incorporates
reliability). The splice is completed by inverting each of the nodes
in the Rival Orb (ie negating parity) and placing them into the
appropriate location in the Player Orb.