Improve qwik.php parityOrb() to avoid fudge factor
Recursion in qwik.php parityOrb() exacerbated a problem in the
computation of the weighted parity average. The $rely has a
range = [0,4] to represent a reliability of [0.0,1.0]. Division /4
accounted for this but because of the recursion meant that the
parity of the nodes deep in the recursion were all but ignored.

The issue was improved by dividing by the average reliability (on
each recursion) rather than the maximum value of rely=4. This has
allowed for the removal of the mysterious fudge factor.