improved the operation of Venue->rename()
Now, after a rename(), there remains a symlink from the OldFile name
to the NewFile name. A cron job should be created to process these
symlinks and find all references to the OldName in Match records
withing the Player XML and change it to the NewName. In the meantime
the change is made on the fly in the Match->contruct()