implement FriendListing.php
The basic idea here is to serve FriendPage.php with an embedded
FriendListing. The page will be cached by a PWA Workbox using the
StaleWhileRevalidate strategy. When FriendPage.php is requested
by a browser then the prior version will be delivered by the cache
then a JSON call is made to friend.listing.json.php to update the
lising on the fly. Meanwhile the cache will be updated to the
current version.